+++ src/slack/test/slack-alert-mocked.test.ts
@@ -7,11 +7,9 @@ import * as slacker from "../slack-alert";
 const base = process.env.PWD || ".";
 const vcsRoot: string = "github";
 const ciProvider: string = "circleci";
-const reportDirectory: string = base + "/src/slack/test/jsonTestPass";
-const videoDirectory: string = base + "/src/slack/test/videosDirPopulated";
-const screenshotDirectory: string =
-  base + "/src/slack/test/screenshotDirPopulated";
-const logger: boolean = false;
+const reportDir: string = base + "/src/slack/test/jsonTestPass";
+const videoDir: string = base + "/src/slack/test/videosDirPopulated";
+const screenshotDir: string = base + "/src/slack/test/screenshotDirPopulated";
 const mock: SlackMock.Instance = SlackMock.SlackMocker({ logLevel: "debug" });
 const mockedHooks = mock.incomingWebhooks;
 
@@ -36,16 +34,13 @@ describe("tester", () => {
   setup();
 
   it("can call a mock slack instance vcs root bitbucket", async () => {
-    const _vcsRoot = "bitbucket";
-
-    await slacker.slackRunner(
+    await slacker.slackRunner({
       ciProvider,
-      _vcsRoot,
-      reportDirectory,
-      videoDirectory,
-      screenshotDirectory,
-      logger
-    );
+      vcsRoot: "bitbucket",
+      reportDir,
+      videoDir,
+      screenshotDir,
+    });
     const body = await returnSlackWebhookCall();
 
     expect(body).toContain("bitbucket");
@@ -56,15 +51,13 @@ describe("tester", () => {
 describe("tester", () => {
   setup();
   it("can provide a simple report with an unknown ci provider", async () => {
-    const _ciProvider = "csdscscsc";
-    await slacker.slackRunner(
-      _ciProvider,
+    await slacker.slackRunner({
+      ciProvider: "csdscscsc",
       vcsRoot,
-      reportDirectory,
-      videoDirectory,
-      screenshotDirectory,
-      logger
-    );
+      reportDir,
+      videoDir,
+      screenshotDir,
+    });
     const body = await returnSlackWebhookCall();
 
     const buildNum = process.env.CIRCLE_BUILD_NUM;
@@ -76,17 +69,33 @@ describe("tester", () => {
 
 describe("tester", () => {
   setup();
-  it("can call a mock slack instance vcs root github", async () => {
-    await slacker.slackRunner(
+  it("can set custom report link when a custom url is provided", async () => {
+    const artifactUrl = "http://example.com";
+    // /report-data.html
+    await slacker.slackRunner({
       ciProvider,
       vcsRoot,
-      reportDirectory,
-      videoDirectory,
-      screenshotDirectory,
-      logger
-    );
+      reportDir: base + "/src/slack/test/jsonTestFail",
+      videoDir,
+      screenshotDir,
+      customUrl: artifactUrl,
+    });
     const body = await returnSlackWebhookCall();
+    expect(body).toContain(`"url":"${artifactUrl}`);
+  });
+});
 
+describe("tester", () => {
+  setup();
+  it("can call a mock slack instance vcs root github", async () => {
+    await slacker.slackRunner({
+      ciProvider,
+      vcsRoot,
+      reportDir,
+      videoDir,
+      screenshotDir,
+    });
+    const body = await returnSlackWebhookCall();
     checkStatus(body, "passed");
     expect(body).toContain("github");
     expect(body).not.toContain("undefined");
@@ -96,15 +105,13 @@ describe("tester", () => {
 describe("tester", () => {
   setup();
   it("can call a mock slack instance vcs root github with a failing test report", async () => {
-    const _reportDirectory: string = base + "/src/slack/test/jsonTestFail";
-    await slacker.slackRunner(
+    await slacker.slackRunner({
       ciProvider,
       vcsRoot,
-      _reportDirectory,
-      videoDirectory,
-      screenshotDirectory,
-      logger
-    );
+      reportDir: base + "/src/slack/test/jsonTestFail",
+      videoDir,
+      screenshotDir,
+    });
     const body = await returnSlackWebhookCall();
 
     checkStatus(body, "failed");
@@ -112,18 +119,138 @@ describe("tester", () => {
     expect(body).not.toContain("undefined");
   });
 });
+
+describe("test custom webhooks per test status", () => {
+  setup();
+  const errorUrl = "https://hooks.slack.com/services/TEA926DBJ/BEBB8FPCL/error";
+  const failedUrl =
+    "https://hooks.slack.com/services/TEA926DBJ/BEBB8FPCL/failed";
+  const passedUrl =
+    "https://hooks.slack.com/services/TEA926DBJ/BEBB8FPCL/passed";
+  const errorUrlMultiple =
+    "https://hooks.slack.com/services/TEA926DBJ/BEBB8FPCL/error1,https://hooks.slack.com/services/TEA926DBJ/BEBB8FPCL/error2";
+  const failedUrlMultiple =
+    "https://hooks.slack.com/services/TEA926DBJ/BEBB8FPCL/failed1,https://hooks.slack.com/services/TEA926DBJ/BEBB8FPCL/failed2";
+  const passedUrlMultiple =
+    "https://hooks.slack.com/services/TEA926DBJ/BEBB8FPCL/passed1,https://hooks.slack.com/services/TEA926DBJ/BEBB8FPCL/passed2";
+
+  it("calls a mock slack instance with specific webhook for failed test runs", async () => {
+    process.env.SLACK_WEBHOOK_FAILED_URL = failedUrl;
+
+    await slacker.slackRunner({
+      ciProvider,
+      vcsRoot,
+      reportDir: base + "/src/slack/test/jsonTestFail",
+      videoDir,
+      screenshotDir,
+    });
+    const body = await returnSlackWebhookCall(failedUrl);
+
+    checkStatus(body, "failed");
+    expect(body).toContain("github");
+    expect(body).not.toContain("undefined");
+    process.env.SLACK_WEBHOOK_FAILED_URL = "";
+  });
+
+  it("calls a mock slack instance with specific webhook for passed test runs", async () => {
+    process.env.SLACK_WEBHOOK_PASSED_URL = passedUrl;
+
+    await slacker.slackRunner({
+      ciProvider,
+      vcsRoot,
+      reportDir: base + "/src/slack/test/jsonTestPass",
+      videoDir,
+      screenshotDir,
+    });
+    const body = await returnSlackWebhookCall(passedUrl);
+    checkStatus(body, "passed");
+    expect(body).toContain("github");
+    expect(body).not.toContain("undefined");
+    expect(mockedHooks.calls).toHaveLength(1);
+    process.env.SLACK_WEBHOOK_PASSED_URL = "";
+  });
+
+  it("calls a mock slack instance with specific webhook for erroring test runs", async () => {
+    process.env.SLACK_WEBHOOK_ERROR_URL = errorUrl;
+
+    await slacker.slackRunner({
+      ciProvider,
+      vcsRoot,
+      reportDir: base + "/src/slack/test/jsonBuildFail",
+      videoDir,
+      screenshotDir,
+    });
+    const body = await returnSlackWebhookCall(errorUrl);
+    checkStatus(body, "build");
+    expect(body).toContain("github");
+    expect(body).not.toContain("undefined");
+    expect(mockedHooks.calls).toHaveLength(1);
+    process.env.SLACK_WEBHOOK_ERROR_URL = "";
+  });
+
+  it("calls a mock slack instance multiple times if more than one webhook is provided", async () => {
+    process.env.SLACK_WEBHOOK_PASSED_URL = passedUrlMultiple;
+
+    await slacker.slackRunner({
+      ciProvider,
+      vcsRoot,
+      reportDir: base + "/src/slack/test/jsonTestPass",
+      videoDir,
+      screenshotDir,
+    });
+    const body = await returnSlackWebhookCall(passedUrlMultiple, 2);
+    checkStatus(body, "passed");
+    expect(body).toContain("github");
+    expect(body).not.toContain("undefined");
+    expect(mockedHooks.calls).toHaveLength(2);
+    process.env.SLACK_WEBHOOK_PASSED_URL = "";
+  });
+});
+
+describe("test onlyFailed flag", () => {
+  setup();
+  it("calls a mock slack instance with failing test report and onlyFailed flag set", async () => {
+    await slacker.slackRunner({
+      ciProvider,
+      vcsRoot,
+      reportDir: base + "/src/slack/test/jsonTestFail",
+      videoDir,
+      screenshotDir,
+      onlyFailed: true,
+    });
+    const body = await returnSlackWebhookCall();
+
+    checkStatus(body, "failed");
+    expect(body).toContain("github");
+    expect(body).not.toContain("undefined");
+  });
+
+  it("does not call a mock slack instance with passing test report and onlyFailed flag set", async () => {
+    const result = await slacker.slackRunner({
+      ciProvider,
+      vcsRoot,
+      reportDir: base + "/src/slack/test/jsonTestPass",
+      videoDir,
+      screenshotDir,
+      onlyFailed: true,
+    });
+    expect(mockedHooks.calls).toHaveLength(0);
+
+    expect(result).toContain(
+      "onlyFailed flag set, test run status was passed, so not sending message"
+    );
+  });
+});
 describe("tester", () => {
   setup();
   it("can call a mock slack instance vcs root github with a failing build report", async () => {
-    const _reportDirectory: string = base + "/src/slack/test/jsonBuildFail";
-    await slacker.slackRunner(
+    await slacker.slackRunner({
       ciProvider,
       vcsRoot,
-      _reportDirectory,
-      videoDirectory,
-      screenshotDirectory,
-      logger
-    );
+      reportDir: base + "/src/slack/test/jsonBuildFail",
+      videoDir,
+      screenshotDir,
+    });
     const body = await returnSlackWebhookCall();
 
     checkStatus(body, "build");
@@ -133,17 +260,15 @@ describe("tester", () => {
 });
 
 describe("Slack Reporter throws error if we cant find the test report", () => {
-  const _reportDirectory: string = "";
   it("throws error when slack isnt available", async () => {
     function mockRunner() {
-      slacker.slackRunner(
+      slacker.slackRunner({
         ciProvider,
         vcsRoot,
-        _reportDirectory,
-        videoDirectory,
-        screenshotDirectory,
-        logger
-      );
+        reportDir: "",
+        videoDir,
+        screenshotDir,
+      });
     }
     expect(mockRunner).toThrowError("Error: Cannot find test report @ ");
   });
@@ -152,15 +277,13 @@ describe("Slack Reporter throws error if we cant find the test report", () => {
 describe("tester", () => {
   setup();
   it("can provide a simple report with an unknown vcsroot provider", async () => {
-    const _vcsRoot = "none";
-    await slacker.slackRunner(
+    await slacker.slackRunner({
       ciProvider,
-      _vcsRoot,
-      reportDirectory,
-      videoDirectory,
-      screenshotDirectory,
-      logger
-    );
+      vcsRoot: "none",
+      reportDir,
+      videoDir,
+      screenshotDir,
+    });
     const body = await returnSlackWebhookCall();
 
     expect(body).not.toContain("commits");
@@ -168,14 +291,25 @@ describe("tester", () => {
   });
 });
 
-function returnSlackWebhookCall() {
+function returnSlackWebhookCall(webhookUrl?: string, noOfCalls?: number) {
   // This checks the slack mock call counter
-  expect(mockedHooks.calls).toHaveLength(1);
+  expect(mockedHooks.calls).toHaveLength(noOfCalls ? noOfCalls : 1);
   // Load the response as json
-  const firstCall = mockedHooks.calls[0];
-  // check our webhook url called in ENV var SLACK_WEBHOOK_URL
-  expect(firstCall.url).toEqual(process.env.SLACK_WEBHOOK_URL);
-  const body = firstCall.params;
+
+  if (!webhookUrl) {
+    webhookUrl = process.env.SLACK_WEBHOOK_URL;
+  }
+  if (webhookUrl) {
+    let i = 0;
+    webhookUrl.split(",").forEach((eachWebhookUrl) => {
+      const firstCall = mockedHooks.calls[i];
+      // check our webhook url called in ENV var SLACK_WEBHOOK_URL
+      expect(firstCall.url).toEqual(eachWebhookUrl);
+      i++;
+    });
+  }
+  const body = mockedHooks.calls[0].params;
+
   return body;
 }
 function messageBuildURL(body: string) {