+++ untyped.d.ts
@@ -1,3 +1,2 @@
-// declare module "slack-mock";
 declare module "cypress-cucumber-preprocessor/steps";
 declare module "cypress-cucumber-preprocessor/lib/tagsHelper";